% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_random_params}
\alias{generate_random_params}
\title{Generate random parameters for transformation.}
\usage{
generate_random_params(params)
}
\arguments{
\item{params}{A number.}
}
\value{
Named list of tranformation parameters.
}
\description{
Generate random parameters for transformation.
}
